+++ packages/dayjs-date-adapter/tests/json-tools.test.ts
@@ -1,0 +1,5 @@
+import './dayjs-test-setup';
+import '@rschedule/dayjs-date-adapter/setup';
+import tests from '../../json-tools/tests/json-tools-test-setup';
+
+tests();