+++ packages/dayjs-date-adapter/tests/rule-tools.test.ts
@@ -1,0 +1,5 @@
+import './dayjs-test-setup';
+import '@rschedule/dayjs-date-adapter/setup';
+import tests from '../../rule-tools/tests/rule-tools-test-setup';
+
+tests();