+++ packages/dayjs-date-adapter/tests/schedule.test.ts
@@ -1,0 +1,5 @@
+import './dayjs-test-setup';
+import '@rschedule/dayjs-date-adapter/setup';
+import tests from '../../core/src/generators/schedule-test-setup';
+
+tests();